Added feature flags and created one for running freeze checks.
