+++ tapir/core/config.py
@@ -1,4 +1,4 @@
-from tapir.core.models import SidebarLinkGroups
+from tapir.core.utils import SidebarLinkGroups
 
 TAPIR_TABLE_CLASSES = "table table-striped table-hover"
 TAPIR_TABLE_TEMPLATE = "django_tables2/bootstrap5.html"
