+++ tapir/shifts/management/commands/run_freeze_checks.py
@@ -2,6 +2,8 @@ import time
 
 from django.core.management.base import BaseCommand
 
+from tapir.core.models import FeatureFlag
+from tapir.shifts.config import FEATURE_FLAG_NAME_FROZEN_MEMBERS
 from tapir.shifts.models import ShiftUserData
 from tapir.shifts.services.frozen_status_service import FrozenStatusService
 
@@ -13,6 +15,9 @@ class Command(BaseCommand):
     )
 
     def handle(self, *args, **options):
+        if not FeatureFlag.get_flag_value(FEATURE_FLAG_NAME_FROZEN_MEMBERS):
+            return
+
         for shift_user_data in ShiftUserData.objects.all():
             if FrozenStatusService.should_freeze_member(shift_user_data):
                 FrozenStatusService.freeze_member_and_send_email(
